[REFACTOR] #39 : 결제 관련 ENUM들 경로 변경으로 인한 File Changed
