+++ src/main/java/com/gdg/z_meet/domain/order/entity/ItemPurchase.java
@@ -1,6 +1,7 @@
 package com.gdg.z_meet.domain.order.entity;
 
 import com.gdg.z_meet.domain.meeting.entity.Team;
+import com.gdg.z_meet.domain.order.entity.enums.ProductType;
 import com.gdg.z_meet.domain.user.entity.User;
 import com.gdg.z_meet.domain.user.entity.UserProfile;
 import com.gdg.z_meet.global.common.BaseEntity;
