+++ src/main/java/com/gdg/z_meet/domain/order/service/approve/KakaoPayApproveService.java
@@ -24,9 +24,8 @@ public class KakaoPayApproveService {
     private final KakaoItemPurchaseRepository itemPurchaseRepository;
     private final KakaoPayIdempotencyService kakaoPayIdempotencyService;
     private final KakaoPayLockService kakaoPayLockService;
-    // private final PaymentCompensationProducer paymentCompensationProducer;
-    // private final KakaoPayCancelService kakaoPayCancelService;
     private final KakaoPayApproveTransactionService transactionService;
+    private final com.gdg.z_meet.domain.order.service.recovery.PaymentRecoveryService paymentRecoveryService;
 
     /**
      * Try-Confirm-Cancel 패턴 적용
@@ -85,8 +84,14 @@ public class KakaoPayApproveService {
             log.error("결제 승인 처리 실패 - orderId: {}, error: {}",
                     parameter.getOrderId(), e.getMessage(), e);
 
-            // 보상 처리 시도
-            // handleCompensation(parameter, e);
+            // 보상 처리 예약 (DB 기반 큐)
+            // TID는 API 호출 성공 후에만 알 수 있는데, 여기서 알 수 없는 경우가 많음.
+            // 하지만 KakaoPayData에 TID가 있다면(API 호출 후 DB 저장 실패 시) 가져올 수 있음.
+            // 여기서는 안전하게 null로 넘기고, Consumer가 조회하도록 함.
+            // 단, API 호출은 성공했으나 DB 저장이 실패한 경우(망취소)를 대비해 TID를 알 수 있다면 넘겨주는 것이 좋음.
+            // 현재 구조상 API 호출 결과(kakaoApiResponse)를 catch 블록에서 알기 어려우므로,
+            // Consumer(Scheduler)가 KakaoPayData를 조회해서 TID가 있으면 취소하는 방식으로 처리.
+            paymentRecoveryService.scheduleRecovery(parameter.getOrderId(), null, "결제 승인 실패: " + e.getMessage());
 
             throw e;
         } finally {
@@ -97,36 +102,6 @@ public class KakaoPayApproveService {
         }
     }
 
-    // private void handleCompensation(KaKaoPayApproveDTO.Parameter parameter, Exception e) {
-    //     try {
-    //         // DB에서 결제 데이터 조회 (트랜잭션이 커밋되었을 가능성이 있으므로 조회 시도)
-    //         // 주의: startPaymentProcessing이 실패했다면 데이터가 없을 수도 있음
-    //         // 따라서 예외 처리를 통해 안전하게 접근
-    //         transactionService.findKakaoPayData(parameter.getOrderId()).ifPresent(kakaoPayData -> {
-    //             String failureStep = determineFailureStep(e);
-    //
-    //             // API 호출 성공 후 내부 처리 실패인 경우 (TID가 있어야 함)
-    //             // 하지만 여기서 TID를 알기 어려우므로, KakaoPayData에 TID가 저장되어 있는지 확인하거나
-    //             // 단순히 실패 상태로만 변경할지 결정해야 함.
-    //             // 현재 구조에서는 API 호출 후 TID를 DB에 저장하는 단계가 completePayment에 있으므로,
-    //             // API 호출은 성공했으나 DB 저장이 안 된 상태일 수 있음.
-    //             // 이 경우 TID를 모르므로 망취소(전체 취소)가 어려울 수 있음.
-    //
-    //             // 만약 API 호출 전 실패라면 단순히 FAILED로 변경
-    //             transactionService.markAsFailed(kakaoPayData.getId(), "결제 승인 실패: " + e.getMessage());
-    //
-    //             // 비동기 보상 메시지 발행 (필요한 경우)
-    //             // TID가 없으면 보상 처리가 제한적일 수 있음
-    //             if (kakaoPayData.getTid() != null) {
-    //                 compensatePaymentAsync(kakaoPayData, kakaoPayData.getTid(),
-    //                         parameter.getUserId(), "내부 처리 실패: " + e.getMessage(), failureStep);
-    //             }
-    //         });
-    //     } catch (Exception ex) {
-    //         log.error("보상 처리 중 오류 발생 - orderId: {}", parameter.getOrderId(), ex);
-    //     }
-    // }
-
     /**
      * 카카오페이 승인 API 호출
      */
@@ -136,46 +111,4 @@ public class KakaoPayApproveService {
                 .requestPaymentApprove(parameter, kakaoPayData)
                 .orElseThrow(() -> new BusinessException(Code.INVALID_KAKAO_API_RESPONSE));
     }
-
-    // /**
-    //  * 비동기 보상 트랜잭션 실행
-    //  */
-    // private void compensatePaymentAsync(KakaoPayData kakaoPayData, String tid, Long userId,
-    //         String cancelReason, String failureStep) {
-    //     try {
-    //         paymentCompensationProducer.sendCompensationMessage(
-    //                 kakaoPayData.getOrderId(),
-    //                 tid,
-    //                 userId,
-    //                 cancelReason,
-    //                 failureStep);
-    //         log.info("보상 트랜잭션 메시지 발행 완료 - orderId: {}", kakaoPayData.getOrderId());
-    //     } catch (Exception e) {
-    //         log.error("보상 트랜잭션 메시지 발행 실패 - orderId: {}, error: {}",
-    //                 kakaoPayData.getOrderId(), e.getMessage(), e);
-    //         // 동기적으로 보상 처리 시도
-    //         try {
-    //             kakaoPayCancelService.compensatePayment(kakaoPayData, cancelReason);
-    //         } catch (Exception ex) {
-    //             log.error("동기 보상 처리도 실패 - orderId: {}", kakaoPayData.getOrderId(), ex);
-    //         }
-    //     }
-    // }
-    //
-    // /**
-    //  * 실패 단계 판단
-    //  */
-    // private String determineFailureStep(Exception e) {
-    //     String message = e.getMessage();
-    //     if (message != null) {
-    //         if (message.contains("PRODUCT_PROCESSING")) {
-    //             return "PRODUCT_PROCESSING";
-    //         } else if (message.contains("ITEM_PURCHASE_CREATION")) {
-    //             return "ITEM_PURCHASE_CREATION";
-    //         } else if (message.contains("STATUS_UPDATE")) {
-    //             return "STATUS_UPDATE";
-    //         }
-    //     }
-    //     return "UNKNOWN";
-    // }
 }
