+++ src/main/java/com/gdg/z_meet/domain/order/service/approve/KakaoPayApproveTransactionService.java
@@ -4,8 +4,8 @@ import com.gdg.z_meet.domain.order.converter.KaKaoPayApproveConverter;
 import com.gdg.z_meet.domain.order.dto.KaKaoPayApproveDTO;
 import com.gdg.z_meet.domain.order.entity.ItemPurchase;
 import com.gdg.z_meet.domain.order.entity.KakaoPayData;
-import com.gdg.z_meet.domain.order.entity.PaymentStatus;
-import com.gdg.z_meet.domain.order.entity.ProductType;
+import com.gdg.z_meet.domain.order.entity.enums.PaymentStatus;
+import com.gdg.z_meet.domain.order.entity.enums.ProductType;
 import com.gdg.z_meet.domain.order.repository.KakaoItemPurchaseRepository;
 import com.gdg.z_meet.domain.order.repository.KakaoPayDataRepository;
 import com.gdg.z_meet.domain.user.entity.User;
