+++ src/main/java/com/gdg/z_meet/domain/order/service/monitoring/KakaoPayLockMonitoringService.java
@@ -1,6 +1,6 @@
 package com.gdg.z_meet.domain.order.service.monitoring;
 
-import com.gdg.z_meet.domain.order.entity.LockEventType;
+import com.gdg.z_meet.domain.order.entity.enums.LockEventType;
 import com.gdg.z_meet.domain.order.entity.LockMonitoring;
 import com.gdg.z_meet.domain.order.repository.LockMonitoringRepository;
 import lombok.RequiredArgsConstructor;
