+++ src/test/java/com/gdg/z_meet/domain/order/integration/KakaoPayApproveIntegrationTest.java
@@ -3,8 +3,8 @@ package com.gdg.z_meet.domain.order.integration;
 import com.gdg.z_meet.domain.order.client.KaKaoPayApiClient;
 import com.gdg.z_meet.domain.order.dto.KaKaoPayApproveDTO;
 import com.gdg.z_meet.domain.order.entity.KakaoPayData;
-import com.gdg.z_meet.domain.order.entity.PaymentStatus;
-import com.gdg.z_meet.domain.order.entity.ProductType;
+import com.gdg.z_meet.domain.order.entity.enums.PaymentStatus;
+import com.gdg.z_meet.domain.order.entity.enums.ProductType;
 import com.gdg.z_meet.domain.order.repository.KakaoItemPurchaseRepository;
 import com.gdg.z_meet.domain.order.repository.KakaoPayDataRepository;
 import com.gdg.z_meet.domain.order.service.approve.KakaoPayApproveService;
